+++ cpp/move.h
@@ -6,13 +6,13 @@
 struct Move {
 
     bool mtype;
-    uint8 ptype, row1, col1, row2, col2;
+    uint ptype, row1, col1, row2, col2;
 
-    Move(uint8);
+    Move(uint);
 
 };
 
-uint8 encode_place(uint8 ptype, uint8 row, uint8 col);
-uint8 encode_move(uint8 row1, uint8 col1, uint8 row2, uint8 col2);
+uint encode_place(uint ptype, uint row, uint col);
+uint encode_move(uint row1, uint col1, uint row2, uint col2);
 
 #endif
